requirements.txt
opencv-python-headless
streamlit
mediapipe
numpy
joblib
Pillow
pyttsx3
tensorflow
scikit-learn
xgboost
tabulate
streamlit-webrtc
av
streamlit-autorefresh
google-generativeai
python-dotenv
